Corrects an implied map() usage as list
In the six package, the map() function returns an iterator instead
of a list. This code was continuing to use the map() return value
as if it were a list and this broke the address_class facts.

This patch changes the code to use the list() method on the return
value of map().